moved projects_models to models_library.projects
