+++ packages/models-library/src/models_library/projects.py
@@ -14,8 +14,17 @@ class ProjectLocked(BaseModel):
     owner: Optional[Owner]
 
 
+class RunningState(str, Enum):
+    not_started = "NOT_STARTED"
+    pending = "PENDING"
+    started = "STARTED"
+    retrying = "RETRY"
+    success = "SUCCESS"
+    failure = "FAILURE"
+
+
 class ProjectRunningState(BaseModel):
-    value: str
+    value: RunningState
 
 
 class ProjectState(BaseModel):
