+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -9,6 +9,12 @@ import aioredlock
 from aiohttp import web
 from jsonschema import ValidationError
 
+from models_library.projects import (
+    Owner,
+    ProjectLocked,
+    ProjectRunningState,
+    ProjectState,
+)
 from servicelib.utils import fire_and_forget_task, logged_gather
 
 from .. import catalog
@@ -21,7 +27,6 @@ from ..users_api import get_user_name
 from . import projects_api
 from .projects_db import APP_PROJECT_DBAPI
 from .projects_exceptions import ProjectInvalidRightsError, ProjectNotFoundError
-from .projects_models import Owner, ProjectLocked, ProjectRunningState, ProjectState
 from .projects_utils import project_uses_available_services
 
 OVERRIDABLE_DOCUMENT_KEYS = [
